src/server/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
anthropic==0.18.1
supabase==2.3.4
qdrant-client==1.7.3
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
python-jose==3.3.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1